+++ CHANGELOG.md
@@ -10,7 +10,7 @@ and this project adheres to
 ### Changed
 - `lak list performance` and `lak info performance` commands will print
 the performance stats till today as opposed to the last saved checkpoint.
-
+- Deleted unnecessary lakshmi.analyze.Analyzer interface.
 
 ## [v2.12.1] - 2024-05-02
 ### Fixed